+++ packages/domain/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./types";
 export * from "./banking";
 export * from "./wallets/invite-tokens";
+export * from "./transactions/dedupe";